tasks

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	printf("Size of int: %zu byte(s)\n", sizeof(int));
-	printf("Size of char: %zu byte(s)\n", sizeof(char));
-	 printf("Size of long: %zu byte(s)\n", sizeof(long));
-	 printf("Size of long long: %zu byte(s)\n", sizeof(long long));
-	printf("Size of float: %zu byte(s)\n", sizeof(float));
+	printf("Size of int: %zu byte(s)\n",(unsigned long)sizeof(int));
+	printf("Size of char: %zu byte(s)\n", (unsigned long)sizeof(char));
+	 printf("Size of long: %zu byte(s)\n", (unsigned long)sizeof(long));
+	 printf("Size of long long: %zu byte(s)\n", (unsigned long)sizeof(long long));
+	printf("Size of float: %zu byte(s)\n", (unsigned long)sizeof(float));
 return (0);
 }
